@@ -20,14 +20,14 @@
 package net.clementraynaud.skoice.bot;
 
 import net.clementraynaud.skoice.Skoice;
-import net.clementraynaud.skoice.api.events.player.PlayerProximityConnectEvent;
-import net.clementraynaud.skoice.api.events.system.SystemReadyEvent;
 import net.clementraynaud.skoice.commands.skoice.arguments.Argument;
 import net.clementraynaud.skoice.lang.DiscordLang;
 import net.clementraynaud.skoice.lang.LangInfo;
 import net.clementraynaud.skoice.listeners.session.ReadyListener;
 import net.clementraynaud.skoice.menus.EmbeddedMenu;
 import net.clementraynaud.skoice.menus.MenuFactory;
+import net.clementraynaud.skoice.model.minecraft.BasePlayer;
+import net.clementraynaud.skoice.model.minecraft.SkoiceCommandSender;
 import net.clementraynaud.skoice.storage.TempYamlFile;
 import net.clementraynaud.skoice.storage.config.ConfigField;
 import net.clementraynaud.skoice.system.ProximityChannel;
@@ -47,9 +47,6 @@ import net.dv8tion.jda.api.entities.channel.concrete.VoiceChannel;
 import net.dv8tion.jda.api.exceptions.ErrorResponseException;
 import net.dv8tion.jda.api.exceptions.InvalidTokenException;
 import net.dv8tion.jda.api.interactions.Interaction;
-import org.bukkit.ChatColor;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -60,6 +57,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
 import java.util.logging.Level;
 import java.util.stream.Collectors;
 
@@ -90,7 +88,7 @@ public class Bot {
         this.connect(null);
     }
 
-    public void connect(CommandSender sender) {
+    public void connect(SkoiceCommandSender sender) {
         if (!this.plugin.getConfigYamlFile().contains(ConfigField.TOKEN.toString())) {
             this.acknowledgeStatus();
             this.plugin.log(Level.WARNING, "logger.warning.no-token");
@@ -98,9 +96,9 @@ public class Bot {
         }
 
         this.plugin.log(Level.INFO, "chat.configuration.bot-connecting");
-        Player tokenManager;
-        if (sender instanceof Player) {
-            tokenManager = (Player) sender;
+        BasePlayer tokenManager;
+        if (sender instanceof BasePlayer) {
+            tokenManager = (BasePlayer) sender;
             this.tokenManagerId = tokenManager.getUniqueId().toString();
             tokenManager.sendMessage(this.plugin.getLang().getMessage("chat.configuration.bot-connecting"));
         } else {
@@ -119,7 +117,7 @@ public class Bot {
         }
 
         byte[] finalBase64TokenBytes = base64TokenBytes;
-        this.plugin.getServer().getScheduler().runTaskAsynchronously(this.plugin, () -> {
+        CompletableFuture.runAsync(() -> {
             try {
                 this.jda = JDABuilder.createDefault(new String(finalBase64TokenBytes))
                         .addEventListeners(new ReadyListener(this.plugin))
@@ -149,7 +147,7 @@ public class Bot {
     public void shutdown() {
         if (this.canShutdown()) {
             if (this.isAvailable()) {
-                new InterruptSystemTask(this.plugin).run();
+                this.runInterruptSystemTask();
             }
 
             this.jda.shutdown();
@@ -161,6 +159,10 @@ public class Bot {
             } catch (InterruptedException ignored) {
             }
         }
+    }
+
+    public void runInterruptSystemTask() {
+        new InterruptSystemTask(this.plugin).run();
     }
 
     public boolean isAdministrator() {
@@ -178,7 +180,7 @@ public class Bot {
 
     public void setDefaultAvatar() {
         if (this.jda.getSelfUser().getDefaultAvatarUrl().equals(this.jda.getSelfUser().getEffectiveAvatarUrl())) {
-            this.plugin.getServer().getScheduler().runTaskAsynchronously(this.plugin, () -> {
+            CompletableFuture.runAsync(() -> {
                 try (InputStream inputStream = new URL("https://clementraynaud.net/Skoice.jpeg").openStream()) {
                     Icon icon = Icon.from(inputStream);
                     this.jda.getSelfUser().getManager().setAvatar(icon).queue();
@@ -193,15 +195,15 @@ public class Bot {
         if (minecraftId == null) {
             new EmbeddedMenu(this).setContent("account-not-linked").message(member.getUser());
         } else {
-            Player player = this.plugin.getServer().getPlayer(UUID.fromString(minecraftId));
+            BasePlayer player = this.plugin.getPlayer(UUID.fromString(minecraftId));
             if (player != null) {
                 player.sendMessage(this.plugin.getLang().getMessage("chat.player.connected"));
-                this.plugin.getServer().getScheduler().runTask(this.plugin, () -> {
-                    PlayerProximityConnectEvent event = new PlayerProximityConnectEvent(minecraftId, member.getId());
-                    this.plugin.getServer().getPluginManager().callEvent(event);
-                });
+                this.callPlayerProximityConnectEvent(minecraftId, member.getId());
             }
         }
+    }
+
+    protected void callPlayerProximityConnectEvent(String minecraftId, String memberId) {
     }
 
     public void retrieveMutedUsers() {
@@ -260,7 +262,7 @@ public class Bot {
             } else {
                 this.guildId = guilds.get(0).getId();
                 this.plugin.getLang().getFormatter().set("guild",
-                        this.getGuild().getName().replace(Character.toString(ChatColor.COLOR_CHAR), ""));
+                        this.getGuild().getName().replace(Character.toString('&'), ""));
 
                 if (this.getGuild().getRequiredMFALevel() == Guild.MFALevel.TWO_FACTOR_AUTH
                         && !this.jda.getSelfUser().isMfaEnabled()) {
@@ -282,10 +284,7 @@ public class Bot {
 
                 } else {
                     this.status = BotStatus.READY;
-                    this.plugin.getServer().getScheduler().runTask(this.plugin, () -> {
-                        SystemReadyEvent event = new SystemReadyEvent();
-                        this.plugin.getServer().getPluginManager().callEvent(event);
-                    });
+                    this.callSystemReadyEvent();
                 }
             }
 
@@ -293,6 +292,9 @@ public class Bot {
         }
 
         this.isStatusAcknowledged = true;
+    }
+
+    protected void callSystemReadyEvent() {
     }
 
     public void updateActivity() {
@@ -304,7 +306,7 @@ public class Bot {
         }
     }
 
-    public void sendIncompleteConfigurationAlert(Player player, boolean sendIfPermissionMissing, boolean force) {
+    public void sendIncompleteConfigurationAlert(BasePlayer player, boolean sendIfPermissionMissing, boolean force) {
         if (!this.isStatusAcknowledged) {
             return;
         }
@@ -312,9 +314,10 @@ public class Bot {
         if (player.hasPermission(Argument.MANAGE_PERMISSION) || force) {
             if (this.status == BotStatus.NOT_CONNECTED) {
                 if (this.plugin.getConfigYamlFile().getBoolean(ConfigField.TOOLTIPS.toString())) {
-                    this.plugin.getServer().getScheduler().runTaskAsynchronously(this.plugin, () ->
-                            this.plugin.adventure().player(player).sendMessage(this.plugin.getLang()
-                                    .getInteractiveMessage("chat.configuration.incomplete-configuration-operator-interactive")));
+                    CompletableFuture.runAsync(() -> {
+                        player.sendMessage(this.plugin.getLang()
+                                .getInteractiveMessage("chat.configuration.incomplete-configuration-operator-interactive"));
+                    });
                 } else {
                     player.sendMessage(this.plugin.getLang().getMessage("chat.configuration.incomplete-configuration-operator"));
                 }
@@ -328,9 +331,9 @@ public class Bot {
         }
     }
 
-    public void sendNoGuildAlert(Player player) {
+    public void sendNoGuildAlert(BasePlayer player) {
         if (this.plugin.getConfigYamlFile().getBoolean(ConfigField.TOOLTIPS.toString())) {
-            this.plugin.adventure().player(player).sendMessage(
+            player.sendMessage(
                     this.plugin.getLang().getInteractiveMessage("chat.configuration.no-guild-interactive"));
         } else {
             player.sendMessage(this.plugin.getLang().getMessage("chat.configuration.no-guild"));
@@ -360,11 +363,11 @@ public class Bot {
         this.isStatusAcknowledged = true;
     }
 
-    public Player getTokenManager() {
+    public BasePlayer getTokenManager() {
         if (this.tokenManagerId == null) {
             return null;
         }
-        return this.plugin.getServer().getPlayer(UUID.fromString(this.tokenManagerId));
+        return this.plugin.getPlayer(UUID.fromString(this.tokenManagerId));
     }
 
     public Guild getGuild() {
