@@ -21,23 +21,25 @@ package net.clementraynaud.skoice.commands.skoice.arguments;
 
 import net.clementraynaud.skoice.Skoice;
 import net.clementraynaud.skoice.bot.BotStatus;
+import net.clementraynaud.skoice.model.minecraft.BasePlayer;
+import net.clementraynaud.skoice.model.minecraft.SkoiceCommandSender;
 import net.clementraynaud.skoice.storage.config.ConfigField;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
+
+import java.util.concurrent.CompletableFuture;
 
 public class ConfigureArgument extends Argument {
 
-    public ConfigureArgument(Skoice plugin, CommandSender sender) {
+    public ConfigureArgument(Skoice plugin, SkoiceCommandSender sender) {
         super(plugin, sender, ArgumentInfo.CONFIGURE.isAllowedInConsole(), ArgumentInfo.CONFIGURE.isPermissionRequired(), ArgumentInfo.CONFIGURE.isHidden());
     }
 
     @Override
     public void run() {
-        Player player = (Player) this.sender;
+        BasePlayer player = (BasePlayer) this.sender;
         if (super.plugin.getBot().getStatus() == BotStatus.NOT_CONNECTED) {
             if (super.plugin.getConfigYamlFile().getBoolean(ConfigField.TOOLTIPS.toString())) {
-                this.plugin.getServer().getScheduler().runTaskAsynchronously(this.plugin, () -> {
-                    this.plugin.adventure().player(player).sendMessage(this.plugin.getLang()
+                CompletableFuture.runAsync(() -> {
+                    player.sendMessage(this.plugin.getLang()
                             .getInteractiveMessage("chat.configuration.bot-creation-interactive"));
                 });
             } else {
