@@ -21,14 +21,15 @@ package net.clementraynaud.skoice.hooks.discordsrv;
 
 import github.scarsz.discordsrv.DiscordSRV;
 import net.clementraynaud.skoice.Skoice;
+import net.clementraynaud.skoice.platforms.spigot.SkoiceSpigot;
 import net.clementraynaud.skoice.storage.config.ConfigField;
 
 public class DiscordSRVHook {
 
-    private final Skoice plugin;
+    private final SkoiceSpigot plugin;
     private DiscordSRVHookImpl discordSRVHookImpl;
 
-    public DiscordSRVHook(Skoice plugin) {
+    public DiscordSRVHook(SkoiceSpigot plugin) {
         this.plugin = plugin;
     }
 
