@@ -20,6 +20,7 @@
 package net.clementraynaud.skoice.hooks.essentialsx;
 
 import net.clementraynaud.skoice.Skoice;
+import net.clementraynaud.skoice.platforms.spigot.SkoiceSpigot;
 import net.essentialsx.api.v2.events.discordlink.DiscordLinkStatusChangeEvent;
 import net.essentialsx.api.v2.services.discord.DiscordService;
 import net.essentialsx.api.v2.services.discordlink.DiscordLinkService;
@@ -34,21 +35,21 @@ import java.util.UUID;
 
 public class EssentialsXHookImpl implements Listener {
 
-    private final Skoice plugin;
+    private final SkoiceSpigot plugin;
     private DiscordLinkService essentialsLinkApi;
     private DiscordService essentialsDiscordApi;
     private Map<String, String> essentialsLinkedAccounts;
     private boolean synchronizationComplete = false;
 
-    public EssentialsXHookImpl(Skoice plugin) {
+    public EssentialsXHookImpl(SkoiceSpigot plugin) {
         this.plugin = plugin;
-        plugin.getServer().getPluginManager().registerEvents(this, plugin);
+        plugin.getPlugin().getServer().getPluginManager().registerEvents(this, plugin.getPlugin());
     }
 
     public void initialize() {
-        this.essentialsLinkApi = this.plugin.getServer().getServicesManager().load(DiscordLinkService.class);
-        this.essentialsDiscordApi = this.plugin.getServer().getServicesManager().load(DiscordService.class);
-        EssentialsDiscordLink ess = (EssentialsDiscordLink) this.plugin.getServer().getPluginManager().getPlugin("EssentialsDiscordLink");
+        this.essentialsLinkApi = this.plugin.getPlugin().getServer().getServicesManager().load(DiscordLinkService.class);
+        this.essentialsDiscordApi = this.plugin.getPlugin().getServer().getServicesManager().load(DiscordService.class);
+        EssentialsDiscordLink ess = (EssentialsDiscordLink) this.plugin.getPlugin().getServer().getPluginManager().getPlugin("EssentialsDiscordLink");
         if (ess != null && ess.getAccountStorage() != null) {
             this.essentialsLinkedAccounts = Collections.unmodifiableMap(ess.getAccountStorage().getRawStorageMap());
             this.synchronizeAccountLinks();
@@ -76,11 +77,11 @@ public class EssentialsXHookImpl implements Listener {
         }
         if (event.isLinked()) {
             if (event.getMemberId() != null && event.getUser() != null && event.getUser().getUUID() != null) {
-                Skoice.api().linkUser(event.getUser().getUUID(), event.getMemberId());
+                SkoiceSpigot.api().linkUser(event.getUser().getUUID(), event.getMemberId());
             }
         } else {
             if (event.getUser() != null && event.getUser().getUUID() != null) {
-                Skoice.api().unlinkUser(event.getUser().getUUID());
+                SkoiceSpigot.api().unlinkUser(event.getUser().getUUID());
             }
         }
     }
@@ -91,12 +92,12 @@ public class EssentialsXHookImpl implements Listener {
         }
 
         Map<String, String> existingHookLinks = this.essentialsLinkedAccounts;
-        Map<String, String> existingSkoiceLinks = new HashMap<>(Skoice.api().getLinkedAccounts());
+        Map<String, String> existingSkoiceLinks = new HashMap<>(SkoiceSpigot.api().getLinkedAccounts());
 
         existingHookLinks.forEach((minecraftId, discordId) -> {
             if (!existingSkoiceLinks.containsValue(discordId)) {
                 try {
-                    Skoice.api().linkUser(UUID.fromString(minecraftId), discordId);
+                    SkoiceSpigot.api().linkUser(UUID.fromString(minecraftId), discordId);
                 } catch (IllegalArgumentException ignored) {
                 }
             }
