@@ -21,7 +21,7 @@ package net.clementraynaud.skoice.lang;
 
 import net.clementraynaud.skoice.util.ConfigurationUtil;
 import net.clementraynaud.skoice.util.MapUtil;
-import org.bukkit.configuration.file.YamlConfiguration;
+import org.simpleyaml.configuration.file.YamlConfiguration;
 
 import java.util.HashMap;
 import java.util.Map;
