@@ -20,51 +20,47 @@
 package net.clementraynaud.skoice.system;
 
 import net.clementraynaud.skoice.Skoice;
+import net.clementraynaud.skoice.model.minecraft.FullPlayer;
+import net.clementraynaud.skoice.model.minecraft.SkoiceGameMode;
 import net.clementraynaud.skoice.storage.config.ConfigField;
 import net.clementraynaud.skoice.util.DistanceUtil;
-import net.clementraynaud.skoice.util.ThreadUtil;
 import net.dv8tion.jda.api.entities.channel.concrete.VoiceChannel;
-import net.kyori.adventure.text.Component;
-import org.bukkit.ChatColor;
-import org.bukkit.GameMode;
-import org.bukkit.entity.Player;
-import org.bukkit.scoreboard.Scoreboard;
-import org.bukkit.scoreboard.Team;
 
 import java.util.EnumSet;
+import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 
-public class LinkedPlayer {
+public final class LinkedPlayer {
 
     private static final int FALLOFF = 3;
 
     private static final Set<LinkedPlayer> onlineLinkedPlayers = ConcurrentHashMap.newKeySet();
-
-    private final Skoice plugin;
-    private final Player player;
-    private final String discordId;
     private final EnumSet<ActionBarAlert> alerts = EnumSet.noneOf(ActionBarAlert.class);
 
-    public LinkedPlayer(Skoice plugin, Player player, String discordId) {
+    private final Skoice plugin;
+    private final FullPlayer player;
+    private final String discordId;
+
+    public LinkedPlayer(Skoice plugin, FullPlayer player, String discordId) {
         this.plugin = plugin;
         this.player = player;
         this.discordId = discordId;
-        LinkedPlayer.onlineLinkedPlayers.add(this);
-    }
-
-    public static Set<LinkedPlayer> getOnlineLinkedPlayers() {
-        ThreadUtil.ensureNotMainThread();
-        return LinkedPlayer.onlineLinkedPlayers;
+        if (this.player != null) {
+            LinkedPlayer.onlineLinkedPlayers.add(this);
+        }
     }
 
     public static void sendActionBarAlerts() {
         LinkedPlayer.onlineLinkedPlayers.forEach(LinkedPlayer::sendActionBarAlert);
     }
 
+    public static Set<LinkedPlayer> getOnlineLinkedPlayers() {
+        return LinkedPlayer.onlineLinkedPlayers;
+    }
+
     public static LinkedPlayer fromMemberId(String memberId) {
-        ThreadUtil.ensureNotMainThread();
         return LinkedPlayer.onlineLinkedPlayers.stream()
                 .filter(p -> p.getDiscordId().equals(memberId))
                 .findFirst().orElse(null);
@@ -72,8 +68,8 @@ public class LinkedPlayer {
 
     public boolean isStateEligible() {
         return (this.plugin.getConfigYamlFile().getBoolean(ConfigField.PLAYERS_ON_DEATH_SCREEN_INCLUDED.toString()) || !this.player.isDead())
-                && (this.plugin.getConfigYamlFile().getBoolean(ConfigField.SPECTATORS_INCLUDED.toString()) || this.player.getGameMode() != GameMode.SPECTATOR)
-                && !this.plugin.getConfigYamlFile().getStringList(ConfigField.DISABLED_WORLDS.toString()).contains(this.player.getWorld().getName());
+                && (this.plugin.getConfigYamlFile().getBoolean(ConfigField.SPECTATORS_INCLUDED.toString()) || this.player.getGameMode() != SkoiceGameMode.SPECTATOR)
+                && !this.plugin.getConfigYamlFile().getStringList(ConfigField.DISABLED_WORLDS.toString()).contains(this.player.getWorld());
     }
 
     public void addActionBarAlert(ActionBarAlert alert) {
@@ -83,18 +79,13 @@ public class LinkedPlayer {
     public void sendActionBarAlert() {
         ActionBarAlert priorityAlert = ActionBarAlert.getPriorityAlert(this.alerts);
         if (priorityAlert != null) {
-            this.plugin.adventure().player(this.player).sendActionBar(
-                    Component.text(ChatColor.translateAlternateColorCodes('&',
-                            this.plugin.getLang().getMessage("action-bar." + priorityAlert)
-                    ))
-            );
+            this.player.sendActionBar(this.plugin.getLang().getMessage("action-bar." + priorityAlert));
         }
 
         this.alerts.clear();
     }
 
     public Set<LinkedPlayer> getPlayersWithinRange() {
-        ThreadUtil.ensureNotMainThread();
         return LinkedPlayer.onlineLinkedPlayers.stream()
                 .filter(p -> p.isInMainVoiceChannel() || p.isInAnyProximityChannel())
                 .filter(p -> !p.equals(this))
@@ -128,18 +119,17 @@ public class LinkedPlayer {
     }
 
     public boolean isCloseEnoughToPlayer(LinkedPlayer linkedPlayer, boolean falloff) {
-        if (!this.player.getWorld().getName().equals(linkedPlayer.player.getWorld().getName())) {
+        if (!this.player.getWorld().equals(linkedPlayer.player.getWorld())) {
             return false;
         }
 
         if (this.plugin.getConfigYamlFile().getBoolean(ConfigField.SEPARATED_TEAMS.toString())) {
-            Scoreboard scoreboard = this.player.getScoreboard();
-            Team playerTeam = scoreboard.getEntryTeam(this.player.getName());
+            String playerTeam = this.player.getTeam();
             if (playerTeam == null) {
-                if (scoreboard.getEntryTeam(linkedPlayer.getBukkitPlayer().getName()) != null) {
+                if (linkedPlayer.getFullPlayer().getTeam() != null) {
                     return false;
                 }
-            } else if (!playerTeam.equals(scoreboard.getEntryTeam(linkedPlayer.getBukkitPlayer().getName()))) {
+            } else if (!playerTeam.equals(linkedPlayer.getFullPlayer().getTeam())) {
                 return false;
             }
         }
@@ -154,11 +144,26 @@ public class LinkedPlayer {
                 && DistanceUtil.getVerticalDistance(this.player.getLocation(), linkedPlayer.player.getLocation()) <= verticalRadius;
     }
 
-    public Player getBukkitPlayer() {
+    public FullPlayer getFullPlayer() {
         return this.player;
     }
 
     public String getDiscordId() {
         return this.discordId;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (o == null || this.getClass() != o.getClass()) {
+            return false;
+        }
+
+        LinkedPlayer that = (LinkedPlayer) o;
+        return Objects.equals(this.player, that.player);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(this.player);
     }
 }
